requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg==3.1.18
numpy==1.26.4
psycopg2-binary==2.9.6
